fodder in chicken coop feeding area lead to instantiation of eggs in random (unoccupied by staticentity [another egg] and non-staticentity/creature). not using math.min() yet to choose the smaller of (1) number of fodder from previous day or (2) number of egg-producing chicken (adult, haven't skipped a day of feeding). also need to implement chicken coop's shipping bin and incubator.

+++ src/edu/pooh/entities/statics/produce_yields/Egg.java
@@ -18,12 +18,12 @@ public class Egg extends StaticEntity
     private int price;
     private boolean inShippingBin;
 
-    public Egg(Handler handler, float x, float y) {
-        super(handler, x, y, Tile.TILE_WIDTH, Tile.TILE_HEIGHT);
+    public Egg(Handler handler, int x, int y) {
+        super(handler, (float)x, (float)y, Tile.TILE_WIDTH, Tile.TILE_HEIGHT);
 
         price = 50;
         inShippingBin = false;
-    } // **** end Egg(Handler, float, float) constructor ****
+    } // **** end Egg(Handler, int, int) constructor ****
 
     @Override
     public int getPrice() {
